requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
websockets==12.0
websocket-client==1.6.4
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-dotenv==1.0.0
python-dateutil==2.8.2